add a FieldNameResolver that caches field names
